[inductor] move dynamic shapes tests into a new file
[ghstack-poisoned]
